[FLINK-15115][kafka] Drop Kafka 0.9
