+++ flink-connectors/flink-sql-connector-kafka-0.11/pom.xml
@@ -60,7 +60,6 @@ under the License.
 								<includes>
 									<include>org.apache.kafka:*</include>
 									<include>org.apache.flink:flink-connector-kafka-base_${scala.binary.version}</include>
-									<include>org.apache.flink:flink-connector-kafka-0.9_${scala.binary.version}</include>
 									<include>org.apache.flink:flink-connector-kafka-0.10_${scala.binary.version}</include>
 									<include>org.apache.flink:flink-connector-kafka-0.11_${scala.binary.version}</include>
 								</includes>
