+++ flink-table/flink-table-common/src/main/java/org/apache/flink/table/descriptors/ConnectorDescriptorValidator.java
@@ -44,7 +44,7 @@ public abstract class ConnectorDescriptorValidator implements DescriptorValidato
 
 	/**
 	 * Key for describing the version of the connector. This property can be used for different
-	 * connector versions (e.g. Kafka 0.9 or Kafka 0.11).
+	 * connector versions (e.g. Kafka 0.10 or Kafka 0.11).
 	 */
 	public static final String CONNECTOR_VERSION = "connector.version";
